restructuring of LazyNutObject and LazyNutObjectModel
- LazyNutObject is a thin wrapper to QDomDocument, offering name, type,
subtype and getValue
- a pointer to LazyNutObject is used in LazyNutObjectCatalogue (now in
enumclasses.h)
- LazyNutObjectModel inherits from LazyNutObject and from
QAbstractItemModel
- LazyNutObjectModel instances are built on the fly when the view in
object explorer needs one, and deleted accordingly
- LazyNutObjectModel presents two columns corresponding to XML label and
value attributes, resp.
- The exception to the above is for a command tag having attribute
expand_to_fill=true. Then no children are shown in the tree, and the
command is composed and shown instead. This is a back-off and a useful
step towards having a view placing a widget (button) to get the missing
data from lazyNut.
- some file renaming has occurred