read-only TrialEditor , XMLForm class hierarchy
- class XMLForm is a recursive renderer of lazyNut XML, aware of lazyNut
tags (string, object, list, command, etc.) and of the main arguments
(label, value).
- class ObjectXML inherits from XMLForm and it's specialised in object
description rendering. At the present state it only overrides the top
"this" element and renders it as a title header
- class TrialXML inherits from ObjectXML and it's specialised in trial
descrription rendering. At the present state it skips "type" and
"subtype" elements and it displays event parameters horizontally instead
of vertically
- this class hierarchy is customisable by subclassing and overriding any
of the virtual build functions.