bug fix: changing trial appends plot form widgets instead of clearing the form
- introduced SessionManager::instance()->setupNoOp(sender()) that e.g.
enables to terminate a job macro after jumping into a next job.
In the present bug fix, the problem was that after a trial run the next
job is a draw, but in case no plot form exists, the draw plot should not
be sent. However, doing nothing stalls the job macro.
setupNoOp allows to send an empty job.