several bug fixes concerning undesired multiple plot get
- eliminated PlotViewer::checkDependencies since recently_modified senses
a modification in plot dependencies. This was sending superfluous plot get
- DataViewer::setCurrentItem checks if the item is valid.
This was not necessary before the fusion of df and plot views. And it was causing
spurious plot gets.
- PlotSettingsBaseWidget::emitValueChanged is back to checking if the value is actually changed.
- currentItemChanged signal is taken direclty from the Ui of a DataViewer
in order to control setForm. The problem is that now dataframeResultsViewer
is hidden, so its visible region is empty (but its ui isn't).
- plot aspect ratio triggers a get only if ratio has change at least roughly 10%.