merge button
- merge() creates a form and populates it in the usual way, so that the lazynut magic
autofilling can work, but it does not show the form.
- not tested enough

- problems arose because at the moment when the merge df is created a signal
triggers a request to show it in Results, but the df description is not there
yet, hence the viewer ignores the signal. This does not happen when
the user clicks Apply in settings widget, since at that time the df description
is already there.